requirements.txt
apache-airflow
requests
snowflake-connector-python
boto3
python-dotenv